+++ argos/model/__init__.py
@@ -6,6 +6,11 @@ from argos.model.mixer import MixerModel
 from argos.model.model import Model
 from argos.model.playback import PlaybackModel
 from argos.model.playlist import PlaylistModel
+from argos.model.random import (
+    RANDOM_TRACKS_CHOICE_STRATEGY,
+    RandomTracksChoice,
+    RandomTracksChoiceState,
+)
 from argos.model.track import TrackModel
 from argos.model.tracklist import TracklistModel, TracklistTrackModel
 from argos.model.utils import PlaybackState
@@ -20,6 +25,9 @@ __all__ = (
     "PlaybackModel",
     "PlaybackState",
     "PlaylistModel",
+    "RandomTracksChoice",
+    "RandomTracksChoiceState",
+    "RANDOM_TRACKS_CHOICE_STRATEGY",
     "TracklistModel",
     "TracklistTrackModel",
     "TrackModel",